+++ src/components/common/LoadingSpinner/index.tsx
@@ -1,5 +1,5 @@
 import SpinnerIcon from '../icons/SpinnerIcon';
-import styles from './css/LoadingSpinner.module.css';
+import styles from './css/index.module.css';
 
 type Props = {
   position?: 'top';
