+++ src/components/common/ResetButton/EliteResetButton.tsx
@@ -5,7 +5,7 @@ import styles from './css/ResetButton.module.css';
 import { resetEliteCollections } from '@/service/request/eliteCollection';
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
-import LoadingSpinner from '../LoadingSpinner/LoadingSpinner';
+import LoadingSpinner from '../LoadingSpinner';
 
 export default function EliteResetButton() {
   const { data: session } = useSession();
